Fix layer violations
